Remove double call of invoked method
We are using QMetaObject and QMetaMethod to verify if a method
exists for the given signature. Since QMetaObject.indexOfMethod()
wants to have the full signature of the member to call, we had
to assemble and normalize the given method/parameters into the
signature. We asume that the method name is encoded in UTF8.

Signed-off-by: Nils Hasenbanck <nils@hasenbanck.de>