workaround for malloc panic when compiled with ming64
This is a workaround for malloc panics when compiling go-qml with mingw64.  I haven't yet dug deep enough into the reasons why the valid looking static lazy singleton is causing the panic (not sure if the issue is Qt, mingw or cgo) but these changes work around the problem.  Also dumped the glib dependency.