Fixed memory leak in imageprovider.
Qt uses "implicit shared memory" for many of its types which is a scheme
similar to copy on write, and is tracked by reference counting in the
constructors/destructor of its classes.

Unfortunately the imageprovider code creates an *QImage, which since it
wasn't delete'd the destructor is not called, so when QML is done with its
own copies, the shared image data can't be freed.